Restructure props/ to match events/ meta-plugin pattern
Move the props-core plugin definition from `props/index.js` to
`props/base.js` (mirroring `events/base.js`). The new `props/index.js`
is a meta-plugin that bundles base + propschange:

    import props from "./base.js";
    import propschange from "./propschange.js";
    export { props, propschange };
    export default { dependencies: [props, propschange] };

Internal imports updated to use `./base.js` where they want props core
(propschange, events/onprops, events/propchange). The top-level
`plugins/index.js` drops its separate `propschange` import — the props
meta-plugin now carries it.

Tests use `props/base.js` where they exercise props core alone (the
`Behavior` suite, `install`, `observed-attributes`, `inheritance`, and
the `events/propchange` test that loads `eventsPlugin` separately). The
propschange test continues to explicitly import both `base` and
`propschange` to document what it exercises.
