+++ test/plugins/props/install.js
@@ -1,5 +1,5 @@
 import { defineElement } from "../../util/dom.js";
-import { default as propsPlugin } from "../../../src/plugins/props/index.js";
+import { default as propsPlugin } from "../../../src/plugins/props/base.js";
 
 export default {
 	name: "Installation",
