+++ test/plugins/props/observed-attributes.js
@@ -1,4 +1,4 @@
-import { default as propsPlugin } from "../../../src/plugins/props/index.js";
+import { default as propsPlugin } from "../../../src/plugins/props/base.js";
 import { defineElement } from "../../util/dom.js";
 
 export default {
